python_microservice/requirements.txt
flask==2.3.3
flask-cors==4.0.0
litellm==1.75.2
python-dotenv==1.0.0

